Moved market_app to market and removed the file
